pyproject.toml
[tool.poetry]
name = "ai-assistant"
version = "0.1.0"
description = "The ai-assistant for algorithm learning"
authors = ["ladavilada <lada.vilada@gmail.com>"]
readme = "README.md"
packages = [
    { include = "ai_assistant", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
langchain = "^0.3.19"
langchain-core = "^0.3.37"
langchain-community = "^0.3.18"
langchain-openai = "^0.2.0"
openai = "^1.3.5"
pydantic = "^2.0.0"
python-dotenv = "^1.0.1"
pypdf = "^5.3.1"
tiktoken = "^0.7.0"
pinecone-client = "^3.0.0"
sentence-transformers = "^3.0.0"
torch = ">=2.0.0,<3.0.0"
chromadb = "^0.5.0"
beautifulsoup4 = "^4.12.2"  # Fixed 'bs4' to 'beautifulsoup4'
firecrawl-py = "^0.0.13"
tavily-python = "^0.3.3"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
pdfminer-six = "^20240706"
langchain-text-splitters = ">=0.3.6,<0.4.0"
langgraph = { extras = ["sqlite"], version = "^0.3.3" }
python-telegram-bot = "^21.11.1"

[tool.poetry.group.dev.dependencies]
langchain-cli = "^0.0.35"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
